Extract score calculation and menus to new files
This removes the final dependency to functions.cpp in tests.
functions.h is now included only by the main application,
though some of its functions should still be extracted and tested.
