+++ src/functions.h
@@ -2,11 +2,6 @@
 
 #include "voting_round.h"
 
-/* -------------- Menu strings -------------- */
-auto activeMenuString(std::optional<VotingRound> const& voting_round, bool const show_intro_screen) -> std::string;
-auto helpString() -> std::string;
-auto newRoundFormatString(uint32_t const number_of_items) -> std::string;
-
 /* -------------- Command line inputs -------------- */
 auto getKey() -> char;
 auto getConfirmation() -> bool;
