+++ src/score_helpers.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include "score_helpers.h"
 
 #include <algorithm>
@@ -11,34 +9,6 @@
 namespace
 {
 
-void incrementWinner(Scores& scores, Item const& item) {
-	auto it = std::find_if(scores.begin(), scores.end(), [&](Score const& score) {
-		return score.item == item;
-		});
-	if (it == scores.end()) {
-		Score score{};
-		score.item = item;
-		score.wins = 1;
-		scores.emplace_back(score);
-	}
-	else {
-		it->wins++;
-	}
-}
-void incrementLoser(Scores& scores, Item const& item) {
-	auto it = std::find_if(scores.begin(), scores.end(), [&](Score const& score) {
-		return score.item == item;
-		});
-	if (it == scores.end()) {
-		Score score{};
-		score.item = item;
-		score.losses = 1;
-		scores.emplace_back(score);
-	}
-	else {
-		it->losses++;
-	}
-}
 auto readNextWord(std::stringstream& stream) -> std::string {
 	if (stream.rdbuf()->in_avail() == 0) {
 		printError("No word to read from stream");
@@ -59,20 +29,6 @@ auto parseNextNumber(std::stringstream& stream) -> std::optional<uint32_t> {
 
 } // namespace
 
-auto calculateScores(Items const& items, Votes const& votes) -> Scores {
-	Scores scores{};
-	for (Vote const& vote : votes) {
-		if (vote.winner == Option::A) {
-			incrementWinner(scores, items[vote.a_idx]);
-			incrementLoser(scores, items[vote.b_idx]);
-		}
-		else {
-			incrementWinner(scores, items[vote.b_idx]);
-			incrementLoser(scores, items[vote.a_idx]);
-		}
-	}
-	return scores;
-}
 
 /* -------------- Print scores -------------- */
 auto sortScores(Scores const& scores) -> Scores {
