+++ src/test/test_get_active_menu_string.cpp
@@ -1,4 +1,4 @@
-#include "functions.h"
+#include "menus.h"
 #include "testing.h"
 
 namespace
