+++ src/test/test_item_order_randomized.cpp
@@ -1,5 +1,5 @@
+#include "calculate_scores.h"
 #include "score.h"
-#include "score_helpers.h"
 #include "testing.h"
 #include "voting_round.h"
